Diagnostic Carma Model

This function verifies if the condition of stationarity is satisfied.

Diagnostic.Carma(carma)

Arguments

carma

An object of class yuima.qmle-class where the slot model is a carma process.

Value

Logical variable. If TRUE, Carma is stationary.

Author

YUIMA TEAM

Examples

mod1 <- setCarma(p = 2, q = 1, scale.par = "sig",
          Carma.var = "y")

param1 <- list(a1 = 1.39631, a2 = 0.05029, b0 = 1,
            b1 = 1, sig = 1)
samp1 <- setSampling(Terminal = 100, n = 200)
#> Warning: 
#> YUIMA: 'delta' (re)defined.

set.seed(123)

sim1 <- simulate(mod1, true.parameter = param1,
          sampling = samp1)

est1 <- qmle(sim1, start = param1)
#> 
#> Starting qmle for carma ... 
#> Warning: 
#> YUIMA: Drift and diffusion parameters must be different. Doing
#>                joint estimation, asymptotic theory may not hold true.

Diagnostic.Carma(est1)
#> [1] TRUE