Methods for an Object of Class yuima.law

Methods for an object of class yuima.law.

rand(object, n, param, ...)
dens(object, x, param, log = FALSE, ...)
cdf(object, q, param, ...)
quant(object, p, param, ...)

Arguments

object

an object of class yuima.law

n

sample size of random number to be generated by means of the method rand.

param

vector containing model parameters

...

additional arguments.

x

vector containing values for the evaluation of the density using the method dens.

log

logical variable. If TRUE the method dens returns the log density at x

q

a vector for the evaluation of the cdf.

p

a set of probabilities for the evaluation of the quantile function.

Value

rand

returns a vector of random numbers.

dens

returns the density at x.

cdf

returns the cumulative distribution function at q.

quant

returns the quantile function at p.

Author

YUIMA TEAM

Note

There may be missing information in the description. Please contribute with suggestions and fixings.

Contacts: Lorenzo Mercuri lorenzo.mercuri@unimi.it