Kalman-Bucy Filter

Estimates values of unobserved variables from observed variables in a Linear State Space Model.

kalmanBucyFilter(
  yuima, params, mean_init, vcov_init = NULL, delta.vcov.solve = 0.001,
  are = FALSE, explicit = FALSE, env = globalenv()
)

Arguments

yuima

A yuima object. The class of yuima@model must be yuima.linear_state_space_model.

params

A list of numeric values representing the names and values of parameters.

mean_init

A numeric value representing the initial value of unobserved variables.

vcov_init

A matrix representing the initial variance-covariance value of unobserved variables. This is required if are = FALSE. It can be a numeric of length 1 if the number of observed variables is 1.

delta.vcov.solve

A numeric value representing the step size in solving the mean squared error of the estimator.

are

A logical value. If TRUE, solve the algebraic Riccati equation.

explicit

A logical value. If TRUE, use the explicit formula to solve the filtering equation. The formula is available only if coefficients are time-independent.

env

An environment. The environment in which the model coefficients are evaluated.

Value

A yuima.kalmanBucyFilter object.

model

A yuima.linear_state_space_model object.

mean

A ts object containing the estimated values of unobserved variables.

vcov

An array object containing the estimated mean squared error of the estimator of unobserved variables.

mean.init

A numeric value representing the initial value of unobserved variables.

vcov.init

A matrix representing the initial mean squared error of the estimator of unobserved variables.

delta

A numeric value representing the time step of observations.

data

A yuima.data object.

References

Liptser, R. S., & Shiryaev, A. N. (2001). Statistics of Random Processes: General Theory. Springer.

Author

YUIMA TEAM

Examples

vcov_init <- matrix(0.1)
mean_init <- 0
a <- 1.5
b <- 0.3
c <- 1
sigma <- 0.02

n <- 10^4
h <- 0.001

trueparam <- list(a = a, b = b, c = c, sigma = sigma)
mod <- setModel(drift = c("-a*X", "c*X"),
                diffusion = matrix(c("b", "0", "0", "sigma"), 2, 2),
                solve.variable = c("X", "Y"), state.variable = c("X", "Y"),
                observed.variable = "Y")

samp <- setSampling(delta = h, n = n)
#> Warning: 
#> YUIMA: 'Terminal' (re)defined.
yuima <- simulate(mod, sampling = samp, true.parameter = trueparam)

res <- kalmanBucyFilter(
  yuima, trueparam, mean_init, vcov_init, 0.001,
  are = FALSE, env = globalenv()
)