Phi-divergence test statistic for stochastic differential equations

Phi-divergence test statistic for stochastic differential equations.

phi.test(yuima, H0, H1, phi, print=FALSE,...)

Arguments

yuima

a yuima object.

H0

a named list of parameter under H0.

H1

a named list of parameter under H1.

phi

the phi function to be used in the test. See Details.

print

you can see a progress of the estimation when print is TRUE.

...

passed to qmle function.

Details

phi.test executes a Phi-divergence test. If H1 is not specified this hypothesis is filled with the QMLE estimates.

If phi is missing, then phi(x)=1-x+x*log(x) and the Phi-divergence statistic corresponds to the likelihood ratio test statistic.

Value

ans

an obkect of class phitest.

Author

The YUIMA Project Team

Examples

if (FALSE) { # \dontrun{
model<- setModel(drift="t1*(t2-x)",diffusion="t3")
T<-10
n<-1000
sampling <- setSampling(Terminal=T,n=n)
yuima<-setYuima(model=model, sampling=sampling)

h0 <- list(t1=0.3, t2=1, t3=0.25)
X <- simulate(yuima, xinit=1, true=h0)
h1 <- list(t1=0.3, t2=0.2, t3=0.1)

phi1 <- function(x) 1-x+x*log(x)

phi.test(X, H0=h0, H1=h1,phi=phi1)
phi.test(X, H0=h0, phi=phi1, start=h0, lower=list(t1=0.1, t2=0.1, t3=0.1), 
   upper=list(t1=2,t2=2,t3=2),method="L-BFGS-B")
phi.test(X, H0=h1, phi=phi1, start=h0, lower=list(t1=0.1, t2=0.1, t3=0.1), 
  upper=list(t1=2,t2=2,t3=2),method="L-BFGS-B")
} # }