qgv

Estimate the local Holder exponent with quadratic generalized variations method

qgv(yuima, filter.type = "Daubechies", order = 2, a = NULL)

Arguments

yuima

A yuima object.

filter.type

The filter.type can be set to "Daubechies" or "Classical".

order

The order of the filter a to be chosen

a

Any other filter

Details

Estimation of the Hurst index and the constant of the fractional Ornstein-Uhlenbeck process.

Value

an object of class qgv

References

Brouste, A., Iacus, S.M. (2013) Parameter estimation for the discretely observed fractional Ornstein-Uhlenbeck process and the Yuima R package, Computational Statistics, pp. 1129--1147.

Author

The YUIMA Project Team

See also

Examples

# Estimating both Hurst parameter and diffusion coefficient in fractional Ornstein-Uhlenbeck

model<-setModel(drift="-x*lambda",hurst=NA,diffusion="theta")
#> Warning: 
#> YUIMA: Solution variable (lhs) not specified. Trying to use state variables.
sampling<-setSampling(T=100,n=10000)
#> Warning: 
#> YUIMA: 'delta' (re)defined.
yui1<-simulate(model,true.param=list(theta=1,lambda=4),hurst=0.7,sampling=sampling)
qgv(yui1)
#> 
#> Fractional OU estimation
#>                 hurst      theta
#> Estimate   0.69380862 0.98518322
#> Std. Error 0.01058748 0.06793207

# Estimating Hurst parameter only in diffusion processes

model2<-setModel(drift="-x*lambda",hurst=NA,diffusion="theta*sqrt(x)")
#> Warning: 
#> YUIMA: Solution variable (lhs) not specified. Trying to use state variables.
sampling<-setSampling(T=1,n=10000)
#> Warning: 
#> YUIMA: 'delta' (re)defined.
yui2<-simulate(model2,true.param=list(theta=1,lambda=4),hurst=0.7,sampling=sampling,xinit=10)
qgv(yui2)
#> 
#> Fractional OU estimation
#>                hurst theta * sqrtx
#> Estimate   0.6936123            NA
#> Std. Error        NA            NA