Continuous Autoregressive Moving Average (p, q) Model

'setCarma' describes the following model:

Vt = c0 + sigma (b0 Xt(0) + ... + b(q) Xt(q))

dXt(0) = Xt(1) dt

...

dXt(p-2) = Xt(p-1) dt

dXt(p-1) = (-a(p) Xt(0) - ... - a(1) Xt(p-1))dt + (gamma(0) + gamma(1) Xt(0) + ... + gamma(p) Xt(p-1))dZt

The continuous ARMA process using the state-space representation as in Brockwell (2000) is obtained by choosing:

gamma(0) = 1, gamma(1) = gamma(2) = ... = gamma(p) = 0.

Please refer to the vignettes and the examples or the yuima documentation for details.

setCarma(p,q,loc.par=NULL,scale.par=NULL,ar.par="a",ma.par="b",
lin.par=NULL,Carma.var="v",Latent.var="x",XinExpr=FALSE, Cogarch=FALSE, ...)

Arguments

p

a non-negative integer that indicates the number of the autoregressive coefficients.

q

a non-negative integer that indicates the number of the moving average coefficients.

loc.par

location coefficient. The default value loc.par=NULL implies that c0=0.

scale.par

scale coefficient. The default value scale.par=NULL implies that sigma=1.

ar.par

a character-string that is the label of the autoregressive coefficients. The default Value is ar.par="a".

ma.par

a character-string that is the label of the moving average coefficients. The default Value is ma.par="b".

Carma.var

a character-string that is the label of the observed process. Defaults to "v".

Latent.var

a character-string that is the label of the unobserved process. Defaults to "x".

lin.par

a character-string that is the label of the linear coefficients. If lin.par=NULL, the default, the 'setCarma' builds the CARMA(p, q) model defined as in Brockwell (2000).

XinExpr

a logical variable. The default value XinExpr=FALSE implies that the starting condition for Latent.var is zero. If XinExpr=TRUE, each component of Latent.var has a parameter as a initial value.

Cogarch

a logical variable. The default value Cogarch=FALSE implies that the parameters are specified according to Brockwell (2000).

...

Arguments to be passed to 'setCarma', such as the slots of yuima.model-class

measure

Levy measure of jump variables.

measure.type

type specification for Levy measure.

xinit

a vector of expressions identyfying the starting conditions for CARMA model.

Details

Please refer to the vignettes and the examples or to the yuimadocs package.

An object of yuima.carma-class contains:

info:

It is an object of carma.info-class which is a list of arguments that identifies the carma(p,q) model

and the same slots in an object of yuima.model-class .

Value

model

an object of yuima.carma-class.

Author

The YUIMA Project Team

Note

There may be missing information in the model description. Please contribute with suggestions and fixings.

References

Brockwell, P. (2000) Continuous-time ARMA processes, Stochastic Processes: Theory and Methods. Handbook of Statistics, 19, (C. R. Rao and D. N. Shandhag, eds.) 249-276. North-Holland, Amsterdam.

Examples

# Ex 1. (Continuous ARMA process driven by a Brownian Motion)
# To describe the state-space representation of a CARMA(p=3,q=1) model:
# Vt=c0+alpha0*X0t+alpha1*X1t
# dX0t = X1t*dt 
# dX1t = X2t*dt
# dX2t = (-beta3*X0t-beta2*X1t-beta1*X2t)dt+dWt
# we set
mod1<-setCarma(p=3, 
               q=1, 
               loc.par="c0")
# Look at the model structure by
str(mod1)
#> Formal class 'yuima.carma' [package "yuima"] with 17 slots
#>   ..@ info           :Formal class 'carma.info' [package "yuima"] with 10 slots
#>   .. .. ..@ p         : num 3
#>   .. .. ..@ q         : num 1
#>   .. .. ..@ loc.par   : chr "c0"
#>   .. .. ..@ scale.par : chr(0) 
#>   .. .. ..@ ar.par    : chr "a"
#>   .. .. ..@ ma.par    : chr "b"
#>   .. .. ..@ lin.par   : chr(0) 
#>   .. .. ..@ Carma.var : chr "v"
#>   .. .. ..@ Latent.var: chr "x"
#>   .. .. ..@ XinExpr   : logi FALSE
#>   ..@ drift          :  expression((b0 * x1 + b1 * x2), (x1), (x2)) ...
#>   ..@ diffusion      :List of 4
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((1))
#>   ..@ hurst          : num 0.5
#>   ..@ jump.coeff     : list()
#>   ..@ measure        : list()
#>   ..@ measure.type   : chr(0) 
#>   ..@ parameter      :Formal class 'model.parameter' [package "yuima"] with 7 slots
#>   .. .. ..@ all      : chr [1:6] "b0" "b1" "a3" "a2" ...
#>   .. .. ..@ common   : chr(0) 
#>   .. .. ..@ diffusion: chr(0) 
#>   .. .. ..@ drift    : chr [1:5] "b0" "b1" "a3" "a2" ...
#>   .. .. ..@ jump     : chr(0) 
#>   .. .. ..@ measure  : chr(0) 
#>   .. .. ..@ xinit    : chr "c0"
#>   ..@ state.variable : chr [1:4] "v" "x0" "x1" "x2"
#>   ..@ jump.variable  : chr(0) 
#>   ..@ time.variable  : chr "t"
#>   ..@ noise.number   : int 1
#>   ..@ equation.number: int 4
#>   ..@ dimension      : int [1:6] 6 0 0 5 0 0
#>   ..@ solve.variable : chr [1:4] "v" "x0" "x1" "x2"
#>   ..@ xinit          :  expression((c0), (0), (0)) ...
#>   ..@ J.flag         : logi FALSE

# Ex 2. (General setCarma model driven by a Brownian Motion)
# To describe the model defined as:
# Vt=c0+alpha0*X0t+alpha1*X1t
# dX0t = X1t*dt 
# dX1t = X2t*dt
# dX2t = (-beta3*X0t-beta2*X1t-beta1*X2t)dt+(c0+alpha0*X0t)dWt
# we set 
mod2 <- setCarma(p=3,
                 q=1,
                 loc.par="c0",
                 ma.par="alpha",
                 ar.par="beta",
                 lin.par="alpha")
# Look at the model structure by
str(mod2)
#> Formal class 'yuima.carma' [package "yuima"] with 17 slots
#>   ..@ info           :Formal class 'carma.info' [package "yuima"] with 10 slots
#>   .. .. ..@ p         : num 3
#>   .. .. ..@ q         : num 1
#>   .. .. ..@ loc.par   : chr "c0"
#>   .. .. ..@ scale.par : chr(0) 
#>   .. .. ..@ ar.par    : chr "beta"
#>   .. .. ..@ ma.par    : chr "alpha"
#>   .. .. ..@ lin.par   : chr "alpha"
#>   .. .. ..@ Carma.var : chr "v"
#>   .. .. ..@ Latent.var: chr "x"
#>   .. .. ..@ XinExpr   : logi FALSE
#>   ..@ drift          :  expression((alpha0 * x1 + alpha1 * x2), (x1), (x2)) ...
#>   ..@ diffusion      :List of 4
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((c0))
#>   ..@ hurst          : num 0.5
#>   ..@ jump.coeff     : list()
#>   ..@ measure        : list()
#>   ..@ measure.type   : chr(0) 
#>   ..@ parameter      :Formal class 'model.parameter' [package "yuima"] with 7 slots
#>   .. .. ..@ all      : chr [1:6] "alpha0" "alpha1" "beta3" "beta2" ...
#>   .. .. ..@ common   : chr(0) 
#>   .. .. ..@ diffusion: chr "c0"
#>   .. .. ..@ drift    : chr [1:5] "alpha0" "alpha1" "beta3" "beta2" ...
#>   .. .. ..@ jump     : chr(0) 
#>   .. .. ..@ measure  : chr(0) 
#>   .. .. ..@ xinit    : chr "c0"
#>   ..@ state.variable : chr [1:4] "v" "x0" "x1" "x2"
#>   ..@ jump.variable  : chr(0) 
#>   ..@ time.variable  : chr "t"
#>   ..@ noise.number   : int 1
#>   ..@ equation.number: int 4
#>   ..@ dimension      : int [1:6] 6 0 1 5 0 0
#>   ..@ solve.variable : chr [1:4] "v" "x0" "x1" "x2"
#>   ..@ xinit          :  expression((c0), (0), (0)) ...
#>   ..@ J.flag         : logi FALSE

# Ex 3. (Continuous Arma model driven by a Levy process)
# To specify the CARMA(p=3,q=1) model driven by a Compound Poisson process defined as:
# Vt=c0+alpha0*X0t+alpha1*X1t
# dX0t = X1t*dt 
# dX1t = X2t*dt
# dX2t = (-beta3*X0t-beta2*X1t-beta1*X2t)dt+dzt
# we set the Levy measure as in setModel
mod3 <- setCarma(p=3,
                 q=1,
                 loc.par="c0",
                 measure=list(intensity="1",df=list("dnorm(z, 0, 1)")),
                 measure.type="CP")
# Look at the model structure by
str(mod3)
#> Formal class 'yuima.carma' [package "yuima"] with 17 slots
#>   ..@ info           :Formal class 'carma.info' [package "yuima"] with 10 slots
#>   .. .. ..@ p         : num 3
#>   .. .. ..@ q         : num 1
#>   .. .. ..@ loc.par   : chr "c0"
#>   .. .. ..@ scale.par : chr(0) 
#>   .. .. ..@ ar.par    : chr "a"
#>   .. .. ..@ ma.par    : chr "b"
#>   .. .. ..@ lin.par   : chr(0) 
#>   .. .. ..@ Carma.var : chr "v"
#>   .. .. ..@ Latent.var: chr "x"
#>   .. .. ..@ XinExpr   : logi FALSE
#>   ..@ drift          :  expression((b0 * x1 + b1 * x2), (x1), (x2)) ...
#>   ..@ diffusion      :List of 4
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   ..@ hurst          : num 0.5
#>   ..@ jump.coeff     :List of 4
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((1))
#>   ..@ measure        :List of 2
#>   .. ..$ df       :List of 2
#>   .. .. ..$ func:function (x, mean = 0, sd = 1, log = FALSE)  
#>   .. .. ..$ expr:  expression(dnorm(z, 0, 1))
#>   .. ..$ intensity:  expression(1)
#>   ..@ measure.type   : chr "CP"
#>   ..@ parameter      :Formal class 'model.parameter' [package "yuima"] with 7 slots
#>   .. .. ..@ all      : chr [1:6] "b0" "b1" "a3" "a2" ...
#>   .. .. ..@ common   : chr(0) 
#>   .. .. ..@ diffusion: chr(0) 
#>   .. .. ..@ drift    : chr [1:5] "b0" "b1" "a3" "a2" ...
#>   .. .. ..@ jump     : chr(0) 
#>   .. .. ..@ measure  : chr(0) 
#>   .. .. ..@ xinit    : chr "c0"
#>   ..@ state.variable : chr [1:4] "v" "x0" "x1" "x2"
#>   ..@ jump.variable  : chr "z"
#>   ..@ time.variable  : chr "t"
#>   ..@ noise.number   : num 1
#>   ..@ equation.number: int 4
#>   ..@ dimension      : int [1:6] 6 0 0 5 0 0
#>   ..@ solve.variable : chr [1:4] "v" "x0" "x1" "x2"
#>   ..@ xinit          :  expression((c0), (0), (0)) ...
#>   ..@ J.flag         : logi FALSE

# Ex 4. (General setCarma  model driven by a Levy process)
# Vt=c0+alpha0*X0t+alpha1*X1t
# dX0t = X1t*dt 
# dX1t = X2t*dt
# dX2t = (-beta3*X1t-beta2*X2t-beta1*X3t)dt+(c0+alpha0*X0t)dzt
mod4 <- setCarma(p=3,
                 q=1,
                 loc.par="c0",
                 ma.par="alpha",
                 ar.par="beta",
                 lin.par="alpha",
                 measure=list(intensity="1",df=list("dnorm(z, 0, 1)")),
                 measure.type="CP")
# Look at the model structure by
str(mod4)
#> Formal class 'yuima.carma' [package "yuima"] with 17 slots
#>   ..@ info           :Formal class 'carma.info' [package "yuima"] with 10 slots
#>   .. .. ..@ p         : num 3
#>   .. .. ..@ q         : num 1
#>   .. .. ..@ loc.par   : chr "c0"
#>   .. .. ..@ scale.par : chr(0) 
#>   .. .. ..@ ar.par    : chr "beta"
#>   .. .. ..@ ma.par    : chr "alpha"
#>   .. .. ..@ lin.par   : chr "alpha"
#>   .. .. ..@ Carma.var : chr "v"
#>   .. .. ..@ Latent.var: chr "x"
#>   .. .. ..@ XinExpr   : logi FALSE
#>   ..@ drift          :  expression((alpha0 * x1 + alpha1 * x2), (x1), (x2)) ...
#>   ..@ diffusion      :List of 4
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   ..@ hurst          : num 0.5
#>   ..@ jump.coeff     :List of 4
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((c0))
#>   ..@ measure        :List of 2
#>   .. ..$ df       :List of 2
#>   .. .. ..$ func:function (x, mean = 0, sd = 1, log = FALSE)  
#>   .. .. ..$ expr:  expression(dnorm(z, 0, 1))
#>   .. ..$ intensity:  expression(1)
#>   ..@ measure.type   : chr "CP"
#>   ..@ parameter      :Formal class 'model.parameter' [package "yuima"] with 7 slots
#>   .. .. ..@ all      : chr [1:6] "alpha0" "alpha1" "beta3" "beta2" ...
#>   .. .. ..@ common   : chr(0) 
#>   .. .. ..@ diffusion: chr(0) 
#>   .. .. ..@ drift    : chr [1:5] "alpha0" "alpha1" "beta3" "beta2" ...
#>   .. .. ..@ jump     : chr "c0"
#>   .. .. ..@ measure  : chr(0) 
#>   .. .. ..@ xinit    : chr "c0"
#>   ..@ state.variable : chr [1:4] "v" "x0" "x1" "x2"
#>   ..@ jump.variable  : chr "z"
#>   ..@ time.variable  : chr "t"
#>   ..@ noise.number   : num 1
#>   ..@ equation.number: int 4
#>   ..@ dimension      : int [1:6] 6 0 0 5 1 0
#>   ..@ solve.variable : chr [1:4] "v" "x0" "x1" "x2"
#>   ..@ xinit          :  expression((c0), (0), (0)) ...
#>   ..@ J.flag         : logi FALSE