Hawkes Process with a Continuous Autoregressive Moving Average(p, q) Intensity

'setCarmaHawkes' describes a self-exciting Hawkes process where the intensity is a CARMA(p,q) process. The model admits the Hawkes process with exponential kernel as a special case but it is able to reproduce a more complex time-dependence structure

setCarmaHawkes(p, q, law = NULL, base.Int = "mu0", 
ar.par = "a", ma.par = "b", Counting.Process = "N", 
Intensity.var = "lambda", Latent.var = "x", time.var = "t", 
Type.Jump = FALSE, XinExpr = FALSE)

Arguments

p

a non-negative integer that indicates the number of the autoregressive coefficients.

q

a non-negative integer that indicates the number of the moving average coefficients.

law

An object of yuima.law-class that describes the jump size. If it is NULL, the jump size is unitary.

base.Int

a character-string that is the label of the baseline Intensity parameter. Defaults to "mu0".

ar.par

a character-string that is the label of the autoregressive coefficients. The default Value is ar.par="a".

ma.par

a character-string that is the label of the moving average coefficients. The default Value is ma.par="b".

Counting.Process

a character-string that is the label of the Counting process. Defaults to "N".

Intensity.var

a character-string that is the label of the Intensity process. Defaults to "lambda"

Latent.var

a character-string that is the label of the unobserved process. Defaults to "x".

time.var

the name of the time variable.

Type.Jump

a logical value. If it is TRUE, the jump size is deterministic.

XinExpr

a logical variable. The default value XinExpr=FALSE implies that the starting condition for Latent.var is zero. If XinExpr=TRUE, each component of Latent.var has a parameter as a initial value.

Value

Model an object of yuima.carmaHawkes-class.

References

Mercuri, L., Perchiazzo, A., & Rroji, E. (2022). A Hawkes model with CARMA (p, q) intensity. Insurance: Mathematics and Economics, 116, 1-26. doi:10.1016/j.insmatheco.2024.01.007 .

Author

The YUIMA Project Team

Contacts: Lorenzo Mercuri lorenzo.mercuri@unimi.it

Examples

if (FALSE) { # \dontrun{
# Definition of an Hawkes with exponential Kernel
mod1 <- setCarmaHawkes(p = 1, q = 0)

# Definition of an Hawkes with a CARMA(2,1) Intensity process
mod2 <- setCarmaHawkes(p = 2, q = 1)
} # }