Continuous-time GARCH (p,q) Process

setCogarch describes the Cogarch(p,q) model introduced in Brockwell et al. (2006):

dGt = sqrt(Vt)dZt

Vt = a0 + (a1 Yt(1) + ... + a(p) Yt(p))

dYt(1) = Yt(2) dt

...

dYt(q-1) = Yt(q) dt

dYt(q) = (-b(q) Yt(1) - ... - b(1) Yt(q))dt + (a0 + (a1 Yt(1) + ... + a(p) Yt(p))d[ZtZt]^{q}

setCogarch(p, q, ar.par = "b", ma.par = "a", loc.par = "a0", Cogarch.var = "g",
   V.var = "v", Latent.var = "y", jump.variable = "z",  time.variable = "t",
   measure = NULL, measure.type = NULL, XinExpr = FALSE, startCogarch = 0,
   work = FALSE, ...)

Arguments

p

a non negative integer that is the number of the moving average coefficients of the Variance process.

q

a non-negative integer that indicates the number of the autoregressive coefficients of the Variance process.

ar.par

a character-string that is the label of the autoregressive coefficients.

ma.par

a character-string that is the label of the autoregressive coefficients.

loc.par

the location coefficient.

Cogarch.var

a character-string that is the label of the observed cogarch process.

V.var

a character-string that is the label of the latent variance process.

Latent.var

a character-string that is the label of the latent process in the state space representation for the variance process.

jump.variable

the jump variable.

time.variable

the time variable.

measure

Levy measure of jump variables.

measure.type

type specification for Levy measure.

XinExpr

a vector of expressions identyfying the starting conditions for Cogarch model.

startCogarch

Start condition for the Cogarch process

work

Internal Variable. In the final release this input will be removed.

...

Arguments to be passed to setCogarch such as the slots of the yuima.model-class

Details

We remark that yuima describes a Cogarch(p,q) model using the formulation proposed in Brockwell et al. (2006). This representation has the Cogarch(1,1) model introduced in Kluppelberg et al. (2004) as a special case. Indeed, by choosing beta = a0 b1, eta = b1 and phi = a1, we obtain the Cogarch(1,1) model proposed in Kluppelberg et al. (2004) defined as the solution of the SDEs:

dGt = sqrt(Vt)dZt

dVt = (beta - eta Vt) dt + phi Vt d[ZtZt]^{q}

Please refer to the vignettes and the examples.

An object of yuima.cogarch-class contains:

info:

It is an object of cogarch.info-class which is a list of arguments that identifies the Cogarch(p,q) model

and the same slots in an object of yuima.model-class .

Value

model

an object of yuima.cogarch-class.

Author

The YUIMA Project Team

Note

There may be missing information in the model description. Please contribute with suggestions and fixings.

References

Brockwell, P., Chadraa, E. and Lindner, A. (2006) Continuous-time GARCH processes, The Annals of Applied Probability, 16, 790-826.

Kluppelberg, C., Lindner, A., and Maller, R. (2004) A continuous-time GARCH process driven by a Levy process: Stationarity and second-order behaviour, Journal of Applied Probability, 41, 601-622.

Stefano M. Iacus, Lorenzo Mercuri, Edit Rroji (2017) COGARCH(p,q): Simulation and Inference with the yuima Package, Journal of Statistical Software, 80(4), 1-49.

Examples

# Ex 1. (Continuous time GARCH process driven by a compound poisson process)
prova<-setCogarch(p=1,q=3,work=FALSE,
                  measure=list(intensity="1", df=list("dnorm(z, 0, 1)")),
                  measure.type="CP",
                  Cogarch.var="y",
                  V.var="v",
                  Latent.var="x")
#> Warning: 
#> YUIMA: Solution variable (lhs) not specified. Trying to use state variables.