Constructor of Hawkes model

'setHawkes' constructs an object of class yuima.Hawkes that is a mathematical description of a multivariate Hawkes model

setHawkes(lower.var = "0", upper.var = "t", var.dt = "s",
  process = "N", dimension = 1, intensity = "lambda",
  ExpKernParm1 = "c", ExpKernParm2 = "a", const = "nu",
  measure = NULL, measure.type = NULL)

Arguments

lower.var

Lower bound in the integral

upper.var

Upper bound in the integral

var.dt

Time variable

process

Counting process

dimension

An integer that indicates the components of the counting process

intensity

Intensity Process

ExpKernParm1

Kernel parameters

ExpKernParm2

Kernel parameters

const

Constant term in the intensity process

measure

Jump size. By default 1

measure.type

Type. By default code.

Details

By default the object is an univariate Hawkes process

Value

The function returns an object of class yuima.Hawkes.

Author

YUIMA Team

Examples

if (FALSE) {
# Definition of an univariate hawkes model

provaHawkes2<-setHawkes()
str(provaHawkes2)

# Simulation

true.par <- list(nu1=0.5, c11=3.5,  a11=4.5)

simprv1 <- simulate(object = provaHawkes2, true.parameter = true.par,
  sampling = setSampling(Terminal =70, n=7000))

plot(simprv1)

# Computation of intensity

lambda1 <- Intensity.PPR(simprv1, param = true.par)

plot(lambda1)

# qmle

res1 <- qmle(simprv1, method="Nelder-Mead", start = true.par)

summary(res1)
}