Creates a "yuima" object by combining "model", "data", "sampling", "characteristic" and "functional"slots.

setYuima constructs an object of yuima-class.

setYuima(data, model, sampling, characteristic, functional)

Arguments

data

an object of yuima.data-class.

model

an object of yuima.model-class.

sampling

an object of yuima.sampling-class.

characteristic

an object of yuima.characteristic-class.

functional

an object of class yuima.functional-class.

Details

The yuima-class object is the main object of the yuima package. Some of the slots can be missing.

The slot data contains the data, either empirical or simulated.

The slot model contains the description of the (statistical) model which is used to generate the data via different simulation schemes, to draw inference from the data or both.

The sampling slot contains information on how the data have been collected or how they should be simulated.

The slot characteristic contains information on PLEASE FINISH THIS. The slot functional contains information on PLEASE FINISH THIS.

Please refer to the vignettes and the examples in the yuimadocs package for more informations.

Value

an object of yuima-class.

Author

The YUIMA Project Team

Examples

# Creation of a yuima object with all slots for a 
# stochastic differential equation
# dXt^e = -theta2 * Xt^e * dt + theta1 * dWt 
diffusion <- matrix(c("theta1"), 1, 1)
drift <- c("-1*theta2*x")
ymodel <- setModel(drift=drift, diffusion=diffusion)
#> Warning: 
#> YUIMA: Solution variable (lhs) not specified. Trying to use state variables.
n <- 100
ysamp <- setSampling(Terminal=1, n=n)
#> Warning: 
#> YUIMA: 'delta' (re)defined.

yuima <- setYuima(model=ymodel, sampling=ysamp)

str(yuima)
#> Formal class 'yuima' [package "yuima"] with 5 slots
#>   ..@ data          :Formal class 'yuima.data' [package "yuima"] with 2 slots
#>   .. .. ..@ original.data: NULL
#>   .. .. ..@ zoo.data     : NULL
#>   ..@ model         :Formal class 'yuima.model' [package "yuima"] with 16 slots
#>   .. .. ..@ drift          :  expression((-1 * theta2 * x))
#>   .. .. ..@ diffusion      :List of 1
#>   .. .. .. ..$ :  expression((theta1))
#>   .. .. ..@ hurst          : num 0.5
#>   .. .. ..@ jump.coeff     : list()
#>   .. .. ..@ measure        : list()
#>   .. .. ..@ measure.type   : chr(0) 
#>   .. .. ..@ parameter      :Formal class 'model.parameter' [package "yuima"] with 7 slots
#>   .. .. .. .. ..@ all      : chr [1:2] "theta2" "theta1"
#>   .. .. .. .. ..@ common   : chr(0) 
#>   .. .. .. .. ..@ diffusion: chr "theta1"
#>   .. .. .. .. ..@ drift    : chr "theta2"
#>   .. .. .. .. ..@ jump     : chr(0) 
#>   .. .. .. .. ..@ measure  : chr(0) 
#>   .. .. .. .. ..@ xinit    : chr(0) 
#>   .. .. ..@ state.variable : chr "x"
#>   .. .. ..@ jump.variable  : chr(0) 
#>   .. .. ..@ time.variable  : chr "t"
#>   .. .. ..@ noise.number   : int 1
#>   .. .. ..@ equation.number: int 1
#>   .. .. ..@ dimension      : int [1:6] 2 0 1 1 0 0
#>   .. .. ..@ solve.variable : chr "x"
#>   .. .. ..@ xinit          :  expression((0))
#>   .. .. ..@ J.flag         : logi FALSE
#>   ..@ sampling      :Formal class 'yuima.sampling' [package "yuima"] with 11 slots
#>   .. .. ..@ Initial      : num 0
#>   .. .. ..@ Terminal     : num 1
#>   .. .. ..@ n            : int 100
#>   .. .. ..@ delta        : num 0.01
#>   .. .. ..@ grid         :List of 1
#>   .. .. .. ..$ : num [1:101] 0 0.01 0.02 0.03 0.04 0.05 0.06 0.07 0.08 0.09 ...
#>   .. .. ..@ random       : logi FALSE
#>   .. .. ..@ regular      : logi TRUE
#>   .. .. ..@ sdelta       : num(0) 
#>   .. .. ..@ sgrid        : num(0) 
#>   .. .. ..@ oindex       : num(0) 
#>   .. .. ..@ interpolation: chr "pt"
#>   ..@ characteristic:Formal class 'yuima.characteristic' [package "yuima"] with 2 slots
#>   .. .. ..@ equation.number: int 1
#>   .. .. ..@ time.scale     : num 1
#>   ..@ functional    :Formal class 'yuima.functional' [package "yuima"] with 4 slots
#>   .. .. ..@ F    : NULL
#>   .. .. ..@ f    : list()
#>   .. .. ..@ xinit: num(0) 
#>   .. .. ..@ e    : num(0)