Calculate the value of functional

Calculate the value of functional associated with sde by Euler scheme.

simFunctional(yuima, expand.var="e")
Fnorm(yuima, expand.var="e")
F0(yuima, expand.var="e")

Arguments

yuima

a yuima object containing model, functional and data.

expand.var

default expand.var="e".

Details

Calculate the value of functional of interest. Fnorm returns normalized one, and F0 returns the value for the case small parameter epsilon = 0. In simFunctional and Fnorm, yuima MUST contains the 'data' slot (X in legacy version)

Value

Fe

a real value

Author

YUIMA Project Team

Note

we need to fix this routine.

Examples

set.seed(123)
# to the Black-Scholes economy:
# dXt^e = Xt^e * dt + e * Xt^e * dWt
diff.matrix <- matrix( c("x*e"), 1,1)
model <- setModel(drift = c("x"), diffusion = diff.matrix)
#> Warning: 
#> YUIMA: Solution variable (lhs) not specified. Trying to use state variables.
# call option is evaluated by averating
# max{ (1/T)*int_0^T Xt^e dt, 0}, the first argument is the functional of interest:
Terminal <- 1
xinit <- c(1)
f <- list( c(expression(x/Terminal)), c(expression(0)))
F <- 0
division <- 1000
e <- .3
samp <- setSampling(Terminal = Terminal, n = division)
#> Warning: 
#> YUIMA: 'delta' (re)defined.
yuima <- setYuima(model = model,sampling = samp) 
yuima <- setFunctional( yuima, xinit=xinit, f=f,F=F,e=e)
# evaluate the functional value

yuima <- simulate(yuima,xinit=xinit,true.par=e)
#> Error in eval(vecV, env): object 'e' not found
Fe <- simFunctional(yuima)
#> Warning: 
#> YUIMA: Length of time series and division do not much.
Fe
#> NULL
Fenorm <- Fnorm(yuima)
#> Warning: 
#> YUIMA: Length of time series and division do not much.
#> Error in eval(f[[r]][k]): object 'Terminal' not found
Fenorm
#> Error in eval(expr, envir, enclos): object 'Fenorm' not found