Additional Methods for LaTeX Representations for Yuima Objects

These methods convert yuima-class, yuima.model-class, yuima.carma-class or yuima.cogarch-class objects to character vectors with LaTeX markup.

# S3 method for class 'yuima'
toLatex(object,...)
# S3 method for class 'yuima.model'
toLatex(object,...)
# S3 method for class 'yuima.carma'
toLatex(object,...)
# S3 method for class 'yuima.cogarch'
toLatex(object,...)

Arguments

object

object of a class yuima, yuima.model or yuima.carma.

...

currently not used.

Details

This method tries to convert a formal description of the model slot of the yuima object into a LaTeX formula. This is just a simple proof of concept and probably further LaTex manipulations for use in papers. Copy and paste of the output of toLatex into a real LaTeX file should do the job.

Examples

# dXt = theta*Xt*dt + dWt
mod1 <- setModel(drift="theta*y", diffusion=1, solve.variable=c("y"))
str(mod1)
#> Formal class 'yuima.model' [package "yuima"] with 16 slots
#>   ..@ drift          :  expression((theta * y))
#>   ..@ diffusion      :List of 1
#>   .. ..$ :  expression((1))
#>   ..@ hurst          : num 0.5
#>   ..@ jump.coeff     : list()
#>   ..@ measure        : list()
#>   ..@ measure.type   : chr(0) 
#>   ..@ parameter      :Formal class 'model.parameter' [package "yuima"] with 7 slots
#>   .. .. ..@ all      : chr "theta"
#>   .. .. ..@ common   : chr(0) 
#>   .. .. ..@ diffusion: chr(0) 
#>   .. .. ..@ drift    : chr "theta"
#>   .. .. ..@ jump     : chr(0) 
#>   .. .. ..@ measure  : chr(0) 
#>   .. .. ..@ xinit    : chr(0) 
#>   ..@ state.variable : chr "x"
#>   ..@ jump.variable  : chr(0) 
#>   ..@ time.variable  : chr "t"
#>   ..@ noise.number   : num 1
#>   ..@ equation.number: int 1
#>   ..@ dimension      : int [1:6] 1 0 0 1 0 0
#>   ..@ solve.variable : chr "y"
#>   ..@ xinit          :  expression((0))
#>   ..@ J.flag         : logi FALSE
toLatex(mod1)
#> %%% Copy and paste the following output in your LaTeX file
#> $$
#>  dy
#>  = 
#>  \theta   \cdot  y dt
#>  +
#>  1
#>  dW1
#> $$
#> $$
#>  y(0)=0
#> $$

# A multi-dimensional (correlated) diffusion process. 
# To describe the following model: 
# X=(X1,X2,X3); dXt = U(t,Xt)dt + V(t)dWt
# For drift coeffcient
U <- c("-x1","-2*x2","-t*x3")
# For diffusion coefficient of X1 
v1 <- function(t) 0.5*sqrt(t)
# For diffusion coefficient of X2
v2 <- function(t) sqrt(t)
# For diffusion coefficient of X3
v3 <- function(t) 2*sqrt(t)
# correlation
rho <- function(t) sqrt(1/2)
# coefficient matrix for diffusion term
V <- matrix( c( "v1(t)",
                "v2(t) * rho(t)",
                "v3(t) * rho(t)",
                "",
                "v2(t) * sqrt(1-rho(t)^2)",
                "",
                "",
                "",
                "v3(t) * sqrt(1-rho(t)^2)" 
               ), 3, 3)
# Model sde using "setModel" function
cor.mod <- setModel(drift = U, diffusion = V,
state.variable=c("x1","x2","x3"), 
solve.variable=c("x1","x2","x3") )
str(cor.mod)
#> Formal class 'yuima.model' [package "yuima"] with 16 slots
#>   ..@ drift          :  expression((-x1), (-2 * x2), (-t * x3))
#>   ..@ diffusion      :List of 3
#>   .. ..$ :  expression((v1(t)), (0), (0))
#>   .. ..$ :  expression((v2(t) * rho(t)), (v2(t) * sqrt(1 - rho(t)^2)), (0))
#>   .. ..$ :  expression((v3(t) * rho(t)), (0), (v3(t) * sqrt(1 - rho(t)^2)))
#>   ..@ hurst          : num 0.5
#>   ..@ jump.coeff     : list()
#>   ..@ measure        : list()
#>   ..@ measure.type   : chr(0) 
#>   ..@ parameter      :Formal class 'model.parameter' [package "yuima"] with 7 slots
#>   .. .. ..@ all      : chr(0) 
#>   .. .. ..@ common   : chr(0) 
#>   .. .. ..@ diffusion: chr(0) 
#>   .. .. ..@ drift    : chr(0) 
#>   .. .. ..@ jump     : chr(0) 
#>   .. .. ..@ measure  : chr(0) 
#>   .. .. ..@ xinit    : chr(0) 
#>   ..@ state.variable : chr [1:3] "x1" "x2" "x3"
#>   ..@ jump.variable  : chr(0) 
#>   ..@ time.variable  : chr "t"
#>   ..@ noise.number   : int 3
#>   ..@ equation.number: int 3
#>   ..@ dimension      : int [1:6] 0 0 0 0 0 0
#>   ..@ solve.variable : chr [1:3] "x1" "x2" "x3"
#>   ..@ xinit          :  expression((0), (0), (0))
#>   ..@ J.flag         : logi FALSE
toLatex(cor.mod)
#> %%% Copy and paste the following output in your LaTeX file
#> $$
#> \left(\begin{array}{c}
#>  dx1  \\  dx2  \\  dx3  \\  
#>  \end{array}\right)
#>  = 
#> \left(\begin{array}{c}
#>  -x1 \\ 
#>  -2  \cdot  x2 \\ 
#>  -t  \cdot  x3 \\ 
#>  \end{array}\right) dt
#>  +
#> \left[\begin{array}{ccc}
#>  v1(t)&0&0 \\ 
#>  v2(t)  \cdot  \rho (t)&v2(t)  \cdot  \sqrt (1 - \rho (t)^2)&0 \\ 
#>  v3(t)  \cdot  \rho (t)&0&v3(t)  \cdot  \sqrt (1 - \rho (t)^2) \\ 
#>  \end{array}\right]
#> \left(\begin{array}{c}
#>  dW1\\ dW2\\ dW3 
#>  \end{array}\right)
#> $$
#> $$
#> \left(\begin{array}{c}
#>  x1(0)=0 \\ 
#>  x2(0)=0 \\ 
#>  x3(0)=0 \\ 
#>  \end{array}\right)
#> $$

# A CARMA(p=3,q=1) process.
carma1<-setCarma(p=3,q=1,loc.par="c",scale.par="s")
str(carma1)
#> Formal class 'yuima.carma' [package "yuima"] with 17 slots
#>   ..@ info           :Formal class 'carma.info' [package "yuima"] with 10 slots
#>   .. .. ..@ p         : num 3
#>   .. .. ..@ q         : num 1
#>   .. .. ..@ loc.par   : chr "c"
#>   .. .. ..@ scale.par : chr "s"
#>   .. .. ..@ ar.par    : chr "a"
#>   .. .. ..@ ma.par    : chr "b"
#>   .. .. ..@ lin.par   : chr(0) 
#>   .. .. ..@ Carma.var : chr "v"
#>   .. .. ..@ Latent.var: chr "x"
#>   .. .. ..@ XinExpr   : logi FALSE
#>   ..@ drift          :  expression((s * b0 * x1 + s * b1 * x2), (x1), (x2)) ...
#>   ..@ diffusion      :List of 4
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((1))
#>   ..@ hurst          : num 0.5
#>   ..@ jump.coeff     : list()
#>   ..@ measure        : list()
#>   ..@ measure.type   : chr(0) 
#>   ..@ parameter      :Formal class 'model.parameter' [package "yuima"] with 7 slots
#>   .. .. ..@ all      : chr [1:7] "s" "b0" "b1" "a3" ...
#>   .. .. ..@ common   : chr(0) 
#>   .. .. ..@ diffusion: chr(0) 
#>   .. .. ..@ drift    : chr [1:6] "s" "b0" "b1" "a3" ...
#>   .. .. ..@ jump     : chr(0) 
#>   .. .. ..@ measure  : chr(0) 
#>   .. .. ..@ xinit    : chr "c"
#>   ..@ state.variable : chr [1:4] "v" "x0" "x1" "x2"
#>   ..@ jump.variable  : chr(0) 
#>   ..@ time.variable  : chr "t"
#>   ..@ noise.number   : int 1
#>   ..@ equation.number: int 4
#>   ..@ dimension      : int [1:6] 7 0 0 6 0 0
#>   ..@ solve.variable : chr [1:4] "v" "x0" "x1" "x2"
#>   ..@ xinit          :  expression((c), (0), (0)) ...
#>   ..@ J.flag         : logi FALSE
toLatex(carma1)
#> %%% Copy and paste the following output in your LaTeX file
#> $$
#> \left\{\begin{array}{l} 
#>  v \left( t \right) =  c + \  s  \cdot  \  b ' x \left( t \right) \\ 
#>  dx \left( t \right) = A x \left( t \right) dt + e dW \left( t \right) \\ 
#>  \end{array}\right.
#> $$
#> $$
#> x ( t ) =\left[ x_0 ( t ) ,\  \ldots \  ,\  x_2 ( t ) \right]'
#> $$
#> $$
#> b = \left[ b_0, \ b_1  ,\  0 \right]'
#> $$
#> $$
#> e = \left[ 0, \ 0, \  1 \right]'
#> $$
#> $$
#> A  = \left[ \begin{array}{ccc}
#>   0 & 1 & 0 \\ 
#>  0 & 0 & 1 \\ 
#>  -a_3 & -a_2 & -a_1 \\ 
#>  \end{array}\right]
#> $$

# A COGARCH(p=3,q=5) process.
cogarch1<-setCogarch(p=3,q=5,
                       measure=list(df=list("rNIG(z, mu00, bu00, 1, 0)")),
                       measure.type="code")
#> Warning: 
#> YUIMA: Solution variable (lhs) not specified. Trying to use state variables.
str(cogarch1)
#> Formal class 'yuima.cogarch' [package "yuima"] with 17 slots
#>   ..@ info           :Formal class 'cogarch.info' [package "yuima"] with 11 slots
#>   .. .. ..@ p           : num 3
#>   .. .. ..@ q           : num 5
#>   .. .. ..@ ar.par      : chr "b"
#>   .. .. ..@ ma.par      : chr "a"
#>   .. .. ..@ loc.par     : chr "a0"
#>   .. .. ..@ Cogarch.var : chr "g"
#>   .. .. ..@ V.var       : chr "v"
#>   .. .. ..@ Latent.var  : chr "y"
#>   .. .. ..@ XinExpr     : logi FALSE
#>   .. .. ..@ measure     :List of 1
#>   .. .. .. ..$ df:List of 1
#>   .. .. .. .. ..$ : chr "rNIG(z, mu00, bu00, 1, 0)"
#>   .. .. ..@ measure.type: chr "code"
#>   ..@ drift          :  expression((0), (a1 * y2 + a2 * y3 + a3 * y4), (y2)) ...
#>   ..@ diffusion      :List of 7
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   .. ..$ :  expression((0))
#>   ..@ hurst          : num 0.5
#>   ..@ jump.coeff     :List of 7
#>   .. ..$ :  expression((sqrt(v)), (0))
#>   .. ..$ :  expression((0), (0))
#>   .. ..$ :  expression((0), (0))
#>   .. ..$ :  expression((0), (0))
#>   .. ..$ :  expression((0), (0))
#>   .. ..$ :  expression((0), (0))
#>   .. ..$ :  expression((0), (a0))
#>   ..@ measure        :List of 1
#>   .. ..$ df:List of 1
#>   .. .. ..$ expr:  expression(rNIG(z, mu00, bu00, 1, 0))
#>   ..@ measure.type   : chr "code"
#>   ..@ parameter      :Formal class 'model.parameter' [package "yuima"] with 7 slots
#>   .. .. ..@ all      : chr [1:11] "a1" "a2" "a3" "b5" ...
#>   .. .. ..@ common   : chr(0) 
#>   .. .. ..@ diffusion: chr(0) 
#>   .. .. ..@ drift    : chr [1:8] "a1" "a2" "a3" "b5" ...
#>   .. .. ..@ jump     : chr "a0"
#>   .. .. ..@ measure  : chr [1:2] "mu00" "bu00"
#>   .. .. ..@ xinit    : chr "a0"
#>   ..@ state.variable : chr [1:7] "g" "v" "y1" "y2" ...
#>   ..@ jump.variable  : chr "z"
#>   ..@ time.variable  : chr "t"
#>   ..@ noise.number   : num 1
#>   ..@ equation.number: int 7
#>   ..@ dimension      : int [1:6] 11 0 0 8 1 2
#>   ..@ solve.variable : chr [1:7] "g" "v" "y1" "y2" ...
#>   ..@ xinit          :  expression((0), (a0), (0)) ...
#>   ..@ J.flag         : logi FALSE
toLatex(cogarch1)
#> %%% Copy and paste the following output in your LaTeX file
#> $$
#> \left\{\begin{array}{l} 
#>  dg \left( t \right) = \ \sqrt { v \left( t \right)} \  dz \left( t \right) \\ 
#>  v \left( t \right) =  a0 + \  a ' y \left( t \right) \\ 
#>  dy \left( t \right) = A y \left( t \right) dt + e d\left[z,z\right]^{q} \left( t \right) \\ 
#>  \end{array}\right.
#> $$
#> $$
#> y ( t ) =\left[ y_1 ( t ) ,\  \ldots \  ,\  y_5 ( t ) \right]'
#> $$
#> $$
#> a = \left[ a_1 ,\  \ldots  \ , \  a_3  ,\  0, \ 0 \right]'
#> $$
#> $$
#> e = \left[ 0, \ \ldots \ , \ 0, \   a0 \right]'
#> $$
#> $$
#> A  = \left[ \begin{array}{cccc}
#>   0 & 1 & \ldots & 0 \\ 
#>  \vdots & \vdots & \ddots & \vdots \\ 
#>  0 & 0 & \ldots & 1 \\ 
#>  -b_5  &   -b_4  & \ldots &  -b_1 \\ 
#>  \end{array}\right]
#> $$