Class for stochastic differential equations

The yuima S4 class is a class of the yuima package.

Slots

data:

an object of class yuima.data-class

model:

an object of class yuima.model-class

sampling:

an object of class yuima.sampling-class

characteristic:

an object of class yuima.characteristic-class

functional:

an object of class yuima.functional-class

Details

The yuima-class object is the main object of the yuima package. Some of the slots may be missing.

The data slot contains the data, either empirical or simulated.

The model contains the description of the (statistical) model which is used to generate the data via different simulation schemes, to draw inference from the data or both.

The sampling slot contains information on how the data have been collected or how they should be generated.

The slot characteristic contains information on PLEASE FINISH THIS. The slot functional contains information on PLEASE FINISH THIS.

Methods

new

signature(x = "yuima", data = "yuima.data", model = "yuima.model", sampling = "yuima.sampling", characteristic = "yuima.characteristic": the function makes a copy of the prototype object from the class definition of yuima-class, then calls the initialize method passing as arguments the newly created object and the remaining arguments.

initialize

signature(x = "yuima", data = "yuima.data", model = "yuima.model", sampling = "yuima.sampling", characteristic = "yuima.characteristic": makes a copy of each argument in the corresponding slots of the object x.

get.data

signature(x = "yuima"): returns the content of the slot data.

plot

signature(x = "yuima", ...): calls plot from the zoo package with argument x@data@zoo.data. Additional arguments ... are passed as is to the plot function.

dim

signature(x = "yuima"): the number of SDEs in the yuima object.

length

signature(x = "yuima"): a vector of length of each SDE described in the yuima object.

cce

signature(x = "yuima"): calculates the asyncronous covariance estimator on the data contained in x@data@zoo.data. For more details see cce.

llag

signature(x = "yuima"): calculates the lead lag estimate r on the data contained in x@data@zoo.data. For more details see llag.

simulate

simulation method. For more information see simulate.

cbind

signature(x = "yuima"): bind yuima.data object.

Author

The YUIMA Project Team