Class for the Mathematical Description of Function of a Stochastic Process

The yuima.Map class is a class of the yuima package that extends the yuima-class it represents a map of a stochastic process

zt = g(theta, Xt, t) : R^{q x d x 1} -> R^{l1 x l2 x ...}

or an operator between two independent stochasic process:

zt = h(theta, Xt, Yt, t)

where Xt and Yt are object of class yuima.model-class or yuima-class with the same dimension.

Slots

Here we report the additional slots of an object of class yuima.Map with respect to the yuima-class:

Output:

It is an object of class info.Map and it is composed by the following slots:

formula:

It is a vector that contains the components of map g(theta, Xt, t) or the operator h(theta, Xt, Yt, t)

dimension:

a numeric object that is the dimensions of the Map.

type:

If type = "Maps", the Map is a map of stochastic process, If type = "Operator", the result is an operator between two independent stochastic process

param

it is an object of class param.Map and it is composed by the following slots:

out.var:

labels for Map.

allparam:

labels of all parameters (model and map/operators).

allparamMap:

labels of map/operator parameters.

common:

common parameters.

Input.var:

labels for inputs.

time.var:

label for time variable.

Methods

simulate

simulation method. For more information see simulate.

% \item{toLatex}{This method converts an object of \code{yuima.poisson-class} to character vectors with LaTeX markup.} % \item{qmle}{Quasi maximum likelihood estimation procedure. For more information see \code{\link{qmle}}. }

Author

The YUIMA Project Team