Class for Quasi Maximum Likelihood Estimation of Point Process Regression Models

The yuima.PPR.qmle class is a class of the yuima package that extends the mle-class of the stats4 package.

Slots

call:

is an object of class language.

coef:

is an object of class numeric that contains estimated parameters.

fullcoef:

is an object of class numeric that contains estimated and fixed parameters.

vcov:

is an object of class matrix.

min:

is an object of class numeric.

minuslogl:

is an object of class function.

method:

is an object of class character.

model:

is an object of class yuima.PPR-class.

Methods

Methods mle

All methods for mle-class are available.

Author

The YUIMA Project Team