Class for the mathematical description of CARMA(p,q) model

The yuima.carma class is a class of the yuima package that extends the yuima.model-class.

Slots

info:

is an carma.info-class object that describes the structure of the CARMA(p,q) model.

drift:

is an R expression which specifies the drift coefficient (a vector).

diffusion:

is an R expression which specifies the diffusion coefficient (a matrix).

hurst:

the Hurst parameter of the gaussian noise. If h=0.5, the process is Wiener otherwise it is fractional Brownian motion with that precise value of the Hurst index. Can be set to NA for further specification.

jump.coeff:

a vector of expressions for the jump component.

measure:

Levy measure for jump variables.

measure.type:

Type specification for Levy measures.

state.variable

a vector of names identifying the names used to denote the state variable in the drift and diffusion specifications.

parameter:

which is a short name for “parameters”, is an object of class model.parameter-class. For more details see model.parameter-class documentation page.

state.variable:

identifies the state variables in the R expression.

jump.variable:

identifies the variable for the jump coefficient.

time.variable:

the time variable.

noise.number:

denotes the number of sources of noise. Currently only for the Gaussian part.

equation.number:

denotes the dimension of the stochastic differential equation.

dimension:

the dimensions of the parameter given in the parameter slot.

solve.variable:

identifies the variable with respect to which the stochastic differential equation has to be solved.

xinit:

contains the initial value of the stochastic differential equation.

J.flag:

wheather jump.coeff include jump.variable.

Methods

simulate

simulation method. For more information see simulate.

toLatex

This method converts an object of yuima.carma-class to character vectors with LaTeX markup.

CarmaNoise

Recovering underlying Levy. For more information see CarmaNoise.

qmle

Quasi maximum likelihood estimation procedure. For more information see qmle.

Author

The YUIMA Project Team