Class "yuima.data" for the data slot of a "yuima" class object

The yuima.data-class is a class of the yuima package used to store the data which are hold in the slot data of an object of the yuima-class.

Objects from this class contain either true data or simulated data.

Slots

original.data:

The original data.

zoo.data:

A list of zoo format data.

Details

Objects in this class are created or initialized using the methods new or initialize or via the function setData. The preferred way to construct an object in this class is to use the function setData.

Objects in this class are used to store the data which are hold in the slot data of an object of the yuima-class.

Objects in this class contain two slots described here.

original.data:

The slot original.data contains, as the name suggests, a copy of the original data passed by the user to methods new or initialize or to the function setData. It is intended for backup purposes.

zoo.data:

When a new object of this class is created or initialized using the original.data, the package tries to convert original.data into an object of class zoo. Once coerced to zoo, the data are stored in the slot zoo.data.

If the conversion fails, the initialization or creation of the object fails.

Internally, the yuima package stores and operates on zoo-type objects.

If data are obtained by simulation, the original.data slot is usually empty.

Methods

new

signature(x = "yuima.data", original.data): the function makes a copy of the prototype object from the class definition of yuima.data-class, then calls the initialize method passing as arguments the newly created object and the original.data.

initialize

signature(x = "yuima.data", original.data): makes a copy of original.data into the slot original.data of x and tries to coerce original.data into an object of class zoo. The result is put in the slot zoo.data of x. If coercion fails, the intialize method fails as well.

get.zoo.data

signature(x = "yuima.data"): returns the content of the slot zoo.data of x.

plot

signature(x = "yuima.data", ...): calls plot from the zoo package with argument x@zoo.data. Additional arguments ... are passed as is to the plot function.

dim

signature(x = "yuima.data"): calls dim from the zoo package with argument x@zoo.data.

length

signature(x = "yuima.data"): calls length from the zoo package with argument x@zoo.data.

cce

signature(x = "yuima.data"): calculates asyncronous covariance estimator on the data contained in x@zoo.data. For more details see cce.

llag

signature(x = "yuima.data"): calculates lead lag estimate on the data contained in x@zoo.data. For more details see llag.

cbind.yuima

signature(x = "yuima.data"): bind yuima.data object.

Author

The YUIMA Project Team