Class for the Mathematical Description of Linear State Space Models

The yuima.linear_state_space_model class is a class in the yuima package for the mathematical description and simulation of linear state space models.

Slots

drift:

An R expression specifying the drift coefficient (a vector).

diffusion:

An R expression specifying the diffusion coefficient (a matrix).

hurst:

The Hurst parameter of the Gaussian noise. If hurst=0.5, the process is a Wiener process; otherwise, it is fractional Brownian motion with the specified Hurst index. Can be set to NA for further specification.

jump.coeff:

A matrix of R expressions for the jump component.

measure:

The Levy measure for jump variables.

measure.type:

The type specification for Levy measures.

state.variable:

A vector of names identifying the state variables used in the drift and diffusion specifications.

parameter:

An object of class model.parameter-class, representing the model parameters. For more details, see the model.parameter-class documentation page.

jump.variable:

The variable for the jump coefficient.

time.variable:

The time variable.

noise.number:

The number of sources of noise, currently only for the Gaussian part.

equation.number:

The dimension of the stochastic differential equation.

dimension:

The dimensions of the parameter given in the parameter slot.

solve.variable:

The variable with respect to which the stochastic differential equation is solved.

xinit:

The initial value of the stochastic differential equation.

J.flag:

Indicates whether jump.coeff includes jump.variable.

is.observed:

Indicates whether each state variable is observed (i.e., data is given) or not.

drift_slope:

An expression specifying the slope of the drift coefficient with respect to unobserved variables (a vector).

drift_intercept:

An expression specifying the intercept of the drift coefficient with respect to unobserved variables (a vector).

Author

The YUIMA Project Team